requirements.txt
Django==3.1.4
django-environ==0.4.5
PyMySQL==1.0.2
mysqlclient==2.0.3
cryptography==3.3.1
djangorestframework==3.12.2
djangorestframework_simplejwt==4.6.0
django-filter==2.4.0
django-redis==4.12.1
requests==2.25.1
flake8==3.8.4
django-cors-headers==3.7.0
PyNaCl==1.4.0
Pillow==8.2.0
PyPDF2==1.26.0
coverage==5.4
tox==3.21.4
black==20.8b1
isort==5.7.0